modify ReadTextListener
