+++ AnkiDroid/src/main/java/com/ichi2/anki/ReadText.java
@@ -18,13 +18,17 @@ package com.ichi2.anki;
 
 import android.content.Context;
 import android.content.res.Resources;
+import android.net.Uri;
 import android.os.Handler;
 import android.speech.tts.TextToSpeech;
+import android.speech.tts.UtteranceProgressListener;
 import android.view.View;
+
 import android.view.WindowManager;
 import android.widget.Toast;
 
 import com.afollestad.materialdialogs.MaterialDialog;
+import com.google.android.material.snackbar.Snackbar;
 import com.ichi2.compat.Compat;
 import com.ichi2.compat.CompatHelper;
 
@@ -47,30 +51,9 @@ public class ReadText {
     private static Compat compat = CompatHelper.getCompat();
     private static Object mTtsParams = compat.initTtsParams();
 
-    public static TextToSpeech getmTts() {
-        return mTts;
-    }
-
-    public static void setmTts(TextToSpeech mTts) {
-        ReadText.mTts = mTts;
-    }
-
-    public static void setmReviewer(WeakReference<Context> mReviewer) {
-        ReadText.mReviewer = mReviewer;
-    }
-
-    public static ArrayList<Locale> getAvailableTtsLocales() {
-        return availableTtsLocales;
-    }
-
-    public static WeakReference<Context> getmReviewer() {
-        return mReviewer;
-    }
-
     public static int getmQuestionAnswer() {
         return mQuestionAnswer;
     }
-
 
     public static void speak(String text, String loc, int queueMode) {
         int result = mTts.setLanguage(localeFromStringIgnoringScriptAndExtensions(loc));
@@ -89,7 +72,7 @@ public class ReadText {
         }
     }
 
-    
+
     public static String getLanguage(long did, int ord, int qa) {
         return MetaDB.getLanguage(mReviewer.get(), did, ord, qa);
     }
@@ -292,6 +275,67 @@ public class ReadText {
                 TextToSpeech.LANG_AVAILABLE;
     }
 
+    public static void initializeTts(Context context, ReadTextListener listener) {
+        // Store weak reference to Activity to prevent memory leak
+        mReviewer = new WeakReference<>(context);
+        // Create new TTS object and setup its onInit Listener
+        mTts = new TextToSpeech(context, new TextToSpeech.OnInitListener() {
+            @Override
+            public void onInit(int status) {
+                if (status == TextToSpeech.SUCCESS) {
+                    // build list of available languages
+                    buildAvailableLanguages();
+                    if (availableTtsLocales.size() > 0) {
+                        // notify the reviewer that TTS has been initialized
+                        Timber.d("TTS initialized and available languages found");
+                        ((AbstractFlashcardViewer) mReviewer.get()).ttsInitialized();
+                    } else {
+                        Toast.makeText(mReviewer.get(), mReviewer.get().getString(R.string.no_tts_available_message), Toast.LENGTH_LONG).show();
+                        Timber.w("TTS initialized but no available languages found");
+                    }
+                    mTts.setOnUtteranceProgressListener(new UtteranceProgressListener() {
+                        @Override
+                        public void onDone(String arg0) {
+                            if (ReadText.sTextQueue.size() > 0) {
+                                String[] text = ReadText.sTextQueue.remove(0);
+                                ReadText.speak(text[0], text[1], TextToSpeech.QUEUE_FLUSH);
+                            }
+                            listener.onDone();
+                        }
+                        @Override
+                        @Deprecated
+                        public void onError(String utteranceId) {
+                            Timber.v("Andoid TTS failed. Check logcat for error. Indicates a problem with Android TTS engine.");
+
+                            final Uri helpUrl = Uri.parse(mReviewer.get().getString(R.string.link_faq_tts));
+                            final AnkiActivity ankiActivity = (AnkiActivity) mReviewer.get();
+                            ankiActivity.mayOpenUrl(helpUrl);
+                            UIUtils.showSnackbar(ankiActivity, R.string.no_tts_available_message, false, R.string.help,
+                                    v -> openTtsHelpUrl(helpUrl), ankiActivity.findViewById(R.id.root_layout),
+                                    new Snackbar.Callback());
+                        }
+                        @Override
+                        public void onStart(String arg0) {
+                            // no nothing
+                        }
+                    });
+                } else {
+                    Toast.makeText(mReviewer.get(), mReviewer.get().getString(R.string.no_tts_available_message), Toast.LENGTH_LONG).show();
+                    Timber.w("TTS not successfully initialized");
+                }
+            }
+        });
+        // Show toast that it's getting initialized, as it can take a while before the sound plays the first time
+        Toast.makeText(context, context.getString(R.string.initializing_tts), Toast.LENGTH_LONG).show();
+    }
+
+
+    private static void openTtsHelpUrl(Uri helpUrl) {
+        AnkiActivity activity =  (AnkiActivity) mReviewer.get();
+        activity.openUrl(helpUrl);
+    }
+
+
     public static void buildAvailableLanguages() {
         availableTtsLocales.clear();
         Locale[] systemLocales = Locale.getAvailableLocales();
@@ -325,5 +369,9 @@ public class ReadText {
             }
             mTts.stop();
         }
+    }
+
+    interface ReadTextListener{
+        public void onDone();
     }
 }
